requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
elevenlabs==2.6.0
executing==2.2.0
fastapi==0.115.6
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
llvmlite==0.44.0
mako==1.3.10
markupsafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.93.2
openai-whisper==20250625
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyaudio==0.2.14
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.19.2
python-dateutil==2.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
aiofiles==24.1.0
